separate ScannerManager
