--- docker-compose.yaml
@@ -1,35 +1,0 @@
-version: "3.3"
-
-services:
-  scanner:
-    container_name: scanner
-    build:
-      context: .
-    volumes:
-      - .:/usr/app
-    ports:
-      - "127.0.0.1:8000:8000"
-    restart: unless-stopped
-    env_file: .env
-    networks:
-      scanner-network:
-    environment:
-      - CONFIG=/usr/app/settings/confing_dev.yaml
-
-  db_2:
-    build:
-      context: ./postgres
-      dockerfile: Dockerfile
-
-    container_name: db_2
-    ports:
-      - "127.0.0.1:5432:5432"
-    env_file: .env
-    volumes:
-      - ./.docker/postgres-data:/var/lib/postgresql/data
-    restart: unless-stopped
-    networks:
-      scanner-network:
-
-networks:
-  scanner-network:
